+++ crates/bevy_internal/src/default_plugins.rs
@@ -40,7 +40,7 @@ plugin_group! {
         bevy_core_pipeline:::CorePipelinePlugin,
         #[cfg(feature = "bevy_render_3d")]
         bevy_render_3d:::MeshPipelinePlugin,
-        #[cfg(feature = "bevy_render_3d")]
+        #[cfg(feature = "bevy_volumetric")]
         bevy_volumetric::plugin:::VolumetricFogPlugin,
         #[cfg(feature = "bevy_anti_aliasing")]
         bevy_anti_aliasing:::AntiAliasingPlugin,
