+++ crates/bevy_render_3d/Cargo.toml
@@ -12,6 +12,7 @@ keywords = ["bevy"]
 std = ["bevy_platform_support/std", "tracing/std"]
 webgl = []
 webgpu = []
+volumetric_light = []
 experimental_pcss = []
 shader_format_glsl = ["bevy_render/shader_format_glsl"]
 trace = ["bevy_render/trace"]
