+++ crates/bevy_volumetric/src/plugin/volumetric_fog.wgsl
