+++ crates/bevy_volumetric/src/render.rs
@@ -2,23 +2,18 @@
 
 use core::array;
 
-use bevy_asset::{weak_handle, AssetId, Handle};
-use bevy_color::ColorToComponents as _;
-use bevy_core_pipeline::{
-    core_3d::Camera3d,
-    prepass::{DeferredPrepass, DepthPrepass, MotionVectorPrepass, NormalPrepass},
-};
+use bevy_asset::AssetId;
 use bevy_derive::{Deref, DerefMut};
 use bevy_ecs::{
     component::Component,
-    entity::Entity,
-    query::{Has, QueryItem, With},
+    query::QueryItem,
     resource::Resource,
-    system::{lifetimeless::Read, Commands, Local, Query, Res, ResMut},
+    system::lifetimeless::Read,
     world::{FromWorld, World},
 };
 use bevy_image::{BevyDefault, Image};
-use bevy_math::{vec4, Mat3A, Mat4, Vec3, Vec3A, Vec4, Vec4Swizzles as _};
+use bevy_math::{Mat4, Vec3, Vec4};
+use bevy_render::render_graph::RenderLabel;
 use bevy_render::{
     mesh::{
         allocator::MeshAllocator, Mesh, MeshVertexBufferLayoutRef, RenderMesh, RenderMeshBufferInfo,
@@ -33,31 +28,29 @@ use bevy_render::{
         BlendOperation, BlendState, CachedRenderPipelineId, ColorTargetState, ColorWrites,
         DynamicBindGroupEntries, DynamicUniformBuffer, Face, FragmentState, LoadOp,
         MultisampleState, Operations, PipelineCache, PrimitiveState, RenderPassColorAttachment,
-        RenderPassDescriptor, RenderPipelineDescriptor, SamplerBindingType, Shader, ShaderStages,
-        ShaderType, SpecializedRenderPipeline, SpecializedRenderPipelines, StoreOp, TextureFormat,
-        TextureSampleType, TextureUsages, VertexState,
+        RenderPassDescriptor, RenderPipelineDescriptor, SamplerBindingType, ShaderStages,
+        ShaderType, SpecializedRenderPipeline, StoreOp, TextureFormat, TextureSampleType,
+        VertexState,
     },
-    renderer::{RenderContext, RenderDevice, RenderQueue},
-    sync_world::RenderEntity,
+    renderer::{RenderContext, RenderDevice},
     texture::GpuImage,
-    view::{ExtractedView, Msaa, ViewDepthTexture, ViewTarget, ViewUniformOffset},
-    Extract,
+    view::{Msaa, ViewDepthTexture, ViewTarget, ViewUniformOffset},
 };
-use bevy_transform::components::GlobalTransform;
+use bevy_render_3d::{
+    MeshPipelineViewLayoutKey, MeshPipelineViewLayouts, MeshViewBindGroup,
+    ViewEnvironmentMapUniformOffset, ViewFogUniformOffset, ViewLightProbesUniformOffset,
+    ViewLightsUniformOffset, ViewScreenSpaceReflectionsUniformOffset,
+};
 use bevy_utils::prelude::default;
+
 use bitflags::bitflags;
 
-use crate::{
-    FogVolume, MeshPipelineViewLayoutKey, MeshPipelineViewLayouts, MeshViewBindGroup,
-    ViewEnvironmentMapUniformOffset, ViewFogUniformOffset, ViewLightProbesUniformOffset,
-    ViewLightsUniformOffset, ViewScreenSpaceReflectionsUniformOffset, VolumetricFog,
-    VolumetricLight,
-};
+use crate::plugin::{CUBE_MESH, PLANE_MESH, VOLUMETRIC_FOG_HANDLE};
 
 bitflags! {
     /// Flags that describe the bind group layout used to render volumetric fog.
     #[derive(Clone, Copy, PartialEq)]
-    struct VolumetricFogBindGroupLayoutKey: u8 {
+    pub struct VolumetricFogBindGroupLayoutKey: u8 {
         /// The framebuffer is multisampled.
         const MULTISAMPLED = 0x1;
         /// The volumetric fog has a 3D voxel density texture.
@@ -69,31 +62,13 @@ bitflags! {
     /// Flags that describe the rasterization pipeline used to render volumetric
     /// fog.
     #[derive(Clone, Copy, PartialEq, Eq, Hash)]
-    struct VolumetricFogPipelineKeyFlags: u8 {
+    pub struct VolumetricFogPipelineKeyFlags: u8 {
         /// The view's color format has high dynamic range.
         const HDR = 0x1;
         /// The volumetric fog has a 3D voxel density texture.
         const DENSITY_TEXTURE = 0x2;
     }
 }
-
-/// The volumetric fog shader.
-pub const VOLUMETRIC_FOG_HANDLE: Handle<Shader> =
-    weak_handle!("481f474c-2024-44bb-8f79-f7c05ced95ea");
-
-/// The plane mesh, which is used to render a fog volume that the camera is
-/// inside.
-///
-/// This mesh is simply stretched to the size of the framebuffer, as when the
-/// camera is inside a fog volume it's essentially a full-screen effect.
-pub const PLANE_MESH: Handle<Mesh> = weak_handle!("92523617-c708-4fd0-b42f-ceb4300c930b");
-
-/// The cube mesh, which is used to render a fog volume that the camera is
-/// outside.
-///
-/// Note that only the front faces of this cuboid will be rasterized in
-/// hardware. The back faces will be calculated in the shader via raytracing.
-pub const CUBE_MESH: Handle<Mesh> = weak_handle!("4a1dd661-2d91-4377-a17a-a914e21e277e");
 
 /// The total number of bind group layouts.
 ///
@@ -102,14 +77,12 @@ pub const CUBE_MESH: Handle<Mesh> = weak_handle!("4a1dd661-2d91-4377-a17a-a914e2
 const VOLUMETRIC_FOG_BIND_GROUP_LAYOUT_COUNT: usize =
     VolumetricFogBindGroupLayoutKey::all().bits() as usize + 1;
 
-/// A matrix that converts from local 1×1×1 space to UVW 3D density texture
-/// space.
-static UVW_FROM_LOCAL: Mat4 = Mat4::from_cols(
-    vec4(1.0, 0.0, 0.0, 0.0),
-    vec4(0.0, 1.0, 0.0, 0.0),
-    vec4(0.0, 0.0, 1.0, 0.0),
-    vec4(0.5, 0.5, 0.5, 1.0),
-);
+/// Render graph nodes specific to 3D PBR rendering.
+#[derive(Debug, Hash, PartialEq, Eq, Clone, RenderLabel)]
+pub enum NodeVolumetric {
+    /// Label for the volumetric lighting pass.
+    VolumetricFog,
+}
 
 /// The GPU pipeline for the volumetric fog postprocessing effect.
 #[derive(Resource)]
@@ -143,71 +116,71 @@ pub struct VolumetricFogNode;
 #[derive(PartialEq, Eq, Hash, Clone)]
 pub struct VolumetricFogPipelineKey {
     /// The layout of the view, which is needed for the raymarching.
-    mesh_pipeline_view_key: MeshPipelineViewLayoutKey,
+    pub mesh_pipeline_view_key: MeshPipelineViewLayoutKey,
 
     /// The vertex buffer layout of the primitive.
     ///
     /// Both planes (used when the camera is inside the fog volume) and cubes
     /// (used when the camera is outside the fog volume) use identical vertex
     /// buffer layouts, so we only need one of them.
-    vertex_buffer_layout: MeshVertexBufferLayoutRef,
+    pub vertex_buffer_layout: MeshVertexBufferLayoutRef,
 
     /// Flags that specify features on the pipeline key.
-    flags: VolumetricFogPipelineKeyFlags,
+    pub flags: VolumetricFogPipelineKeyFlags,
 }
 
-/// The same as [`VolumetricFog`] and [`FogVolume`], but formatted for
+/// The same as [`VolumetricFog`](super::VolumetricFog) and [`FogVolume`](super::FogVolume), but formatted for
 /// the GPU.
 ///
 /// See the documentation of those structures for more information on these
 /// fields.
 #[derive(ShaderType)]
 pub struct VolumetricFogUniform {
-    clip_from_local: Mat4,
+    pub clip_from_local: Mat4,
 
     /// The transform from world space to 3D density texture UVW space.
-    uvw_from_world: Mat4,
+    pub uvw_from_world: Mat4,
 
     /// View-space plane equations of the far faces of the fog volume cuboid.
     ///
     /// The vector takes the form V = (N, -N⋅Q), where N is the normal of the
     /// plane and Q is any point in it, in view space. The equation of the plane
     /// for homogeneous point P = (Px, Py, Pz, Pw) is V⋅P = 0.
-    far_planes: [Vec4; 3],
+    pub far_planes: [Vec4; 3],
 
-    fog_color: Vec3,
-    light_tint: Vec3,
-    ambient_color: Vec3,
-    ambient_intensity: f32,
-    step_count: u32,
+    pub fog_color: Vec3,
+    pub light_tint: Vec3,
+    pub ambient_color: Vec3,
+    pub ambient_intensity: f32,
+    pub step_count: u32,
 
     /// The radius of a sphere that bounds the fog volume in view space.
-    bounding_radius: f32,
+    pub bounding_radius: f32,
 
-    absorption: f32,
-    scattering: f32,
-    density: f32,
-    density_texture_offset: Vec3,
-    scattering_asymmetry: f32,
-    light_intensity: f32,
-    jitter_strength: f32,
+    pub absorption: f32,
+    pub scattering: f32,
+    pub density: f32,
+    pub density_texture_offset: Vec3,
+    pub scattering_asymmetry: f32,
+    pub light_intensity: f32,
+    pub jitter_strength: f32,
 }
 
 /// Specifies the offset within the [`VolumetricFogUniformBuffer`] of the
 /// [`VolumetricFogUniform`] for a specific view.
 #[derive(Component, Deref, DerefMut)]
-pub struct ViewVolumetricFog(Vec<ViewFogVolume>);
+pub struct ViewVolumetricFog(pub Vec<ViewFogVolume>);
 
 /// Information that the render world needs to maintain about each fog volume.
 pub struct ViewFogVolume {
     /// The 3D voxel density texture for this volume, if present.
-    density_texture: Option<AssetId<Image>>,
+    pub density_texture: Option<AssetId<Image>>,
     /// The offset of this view's [`VolumetricFogUniform`] structure within the
     /// [`VolumetricFogUniformBuffer`].
-    uniform_buffer_offset: u32,
+    pub uniform_buffer_offset: u32,
     /// True if the camera is outside the fog volume; false if it's inside the
     /// fog volume.
-    exterior: bool,
+    pub exterior: bool,
 }
 
 /// The GPU buffer that stores the [`VolumetricFogUniform`] data.
@@ -267,50 +240,6 @@ impl FromWorld for VolumetricFogPipeline {
             mesh_view_layouts: mesh_view_layouts.clone(),
             volumetric_view_bind_group_layouts: bind_group_layouts,
         }
-    }
-}
-
-/// Extracts [`VolumetricFog`], [`FogVolume`], and [`VolumetricLight`]s
-/// from the main world to the render world.
-pub fn extract_volumetric_fog(
-    mut commands: Commands,
-    view_targets: Extract<Query<(RenderEntity, &VolumetricFog)>>,
-    fog_volumes: Extract<Query<(RenderEntity, &FogVolume, &GlobalTransform)>>,
-    volumetric_lights: Extract<Query<(RenderEntity, &VolumetricLight)>>,
-) {
-    if volumetric_lights.is_empty() {
-        // TODO: needs better way to handle clean up in render world
-        for (entity, ..) in view_targets.iter() {
-            commands
-                .entity(entity)
-                .remove::<(VolumetricFog, ViewVolumetricFogPipelines, ViewVolumetricFog)>();
-        }
-        for (entity, ..) in fog_volumes.iter() {
-            commands.entity(entity).remove::<FogVolume>();
-        }
-        return;
-    }
-
-    for (entity, volumetric_fog) in view_targets.iter() {
-        commands
-            .get_entity(entity)
-            .expect("Volumetric fog entity wasn't synced.")
-            .insert(*volumetric_fog);
-    }
-
-    for (entity, fog_volume, fog_transform) in fog_volumes.iter() {
-        commands
-            .get_entity(entity)
-            .expect("Fog volume entity wasn't synced.")
-            .insert((*fog_volume).clone())
-            .insert(*fog_transform);
-    }
-
-    for (entity, volumetric_light) in volumetric_lights.iter() {
-        commands
-            .get_entity(entity)
-            .expect("Volumetric light entity wasn't synced.")
-            .insert(*volumetric_light);
     }
 }
 
@@ -608,213 +537,6 @@ impl SpecializedRenderPipeline for VolumetricFogPipeline {
     }
 }
 
-/// Specializes volumetric fog pipelines for all views with that effect enabled.
-pub fn prepare_volumetric_fog_pipelines(
-    mut commands: Commands,
-    pipeline_cache: Res<PipelineCache>,
-    mut pipelines: ResMut<SpecializedRenderPipelines<VolumetricFogPipeline>>,
-    volumetric_lighting_pipeline: Res<VolumetricFogPipeline>,
-    view_targets: Query<
-        (
-            Entity,
-            &ExtractedView,
-            &Msaa,
-            Has<NormalPrepass>,
-            Has<DepthPrepass>,
-            Has<MotionVectorPrepass>,
-            Has<DeferredPrepass>,
-        ),
-        With<VolumetricFog>,
-    >,
-    meshes: Res<RenderAssets<RenderMesh>>,
-) {
-    let Some(plane_mesh) = meshes.get(&PLANE_MESH) else {
-        // There's an off chance that the mesh won't be prepared yet if `RenderAssetBytesPerFrame` limiting is in use.
-        return;
-    };
-
-    for (
-        entity,
-        view,
-        msaa,
-        normal_prepass,
-        depth_prepass,
-        motion_vector_prepass,
-        deferred_prepass,
-    ) in view_targets.iter()
-    {
-        // Create a mesh pipeline view layout key corresponding to the view.
-        let mut mesh_pipeline_view_key = MeshPipelineViewLayoutKey::from(*msaa);
-        mesh_pipeline_view_key.set(MeshPipelineViewLayoutKey::NORMAL_PREPASS, normal_prepass);
-        mesh_pipeline_view_key.set(MeshPipelineViewLayoutKey::DEPTH_PREPASS, depth_prepass);
-        mesh_pipeline_view_key.set(
-            MeshPipelineViewLayoutKey::MOTION_VECTOR_PREPASS,
-            motion_vector_prepass,
-        );
-        mesh_pipeline_view_key.set(
-            MeshPipelineViewLayoutKey::DEFERRED_PREPASS,
-            deferred_prepass,
-        );
-
-        let mut textureless_flags = VolumetricFogPipelineKeyFlags::empty();
-        textureless_flags.set(VolumetricFogPipelineKeyFlags::HDR, view.hdr);
-
-        // Specialize the pipeline.
-        let textureless_pipeline_key = VolumetricFogPipelineKey {
-            mesh_pipeline_view_key,
-            vertex_buffer_layout: plane_mesh.layout.clone(),
-            flags: textureless_flags,
-        };
-        let textureless_pipeline_id = pipelines.specialize(
-            &pipeline_cache,
-            &volumetric_lighting_pipeline,
-            textureless_pipeline_key.clone(),
-        );
-        let textured_pipeline_id = pipelines.specialize(
-            &pipeline_cache,
-            &volumetric_lighting_pipeline,
-            VolumetricFogPipelineKey {
-                flags: textureless_pipeline_key.flags
-                    | VolumetricFogPipelineKeyFlags::DENSITY_TEXTURE,
-                ..textureless_pipeline_key
-            },
-        );
-
-        commands.entity(entity).insert(ViewVolumetricFogPipelines {
-            textureless: textureless_pipeline_id,
-            textured: textured_pipeline_id,
-        });
-    }
-}
-
-/// A system that converts [`VolumetricFog`] into [`VolumetricFogUniform`]s.
-pub fn prepare_volumetric_fog_uniforms(
-    mut commands: Commands,
-    mut volumetric_lighting_uniform_buffer: ResMut<VolumetricFogUniformBuffer>,
-    view_targets: Query<(Entity, &ExtractedView, &VolumetricFog)>,
-    fog_volumes: Query<(Entity, &FogVolume, &GlobalTransform)>,
-    render_device: Res<RenderDevice>,
-    render_queue: Res<RenderQueue>,
-    mut local_from_world_matrices: Local<Vec<Mat4>>,
-) {
-    // Do this up front to avoid O(n^2) matrix inversion.
-    local_from_world_matrices.clear();
-    for (_, _, fog_transform) in fog_volumes.iter() {
-        local_from_world_matrices.push(fog_transform.compute_matrix().inverse());
-    }
-
-    let uniform_count = view_targets.iter().len() * local_from_world_matrices.len();
-
-    let Some(mut writer) =
-        volumetric_lighting_uniform_buffer.get_writer(uniform_count, &render_device, &render_queue)
-    else {
-        return;
-    };
-
-    for (view_entity, extracted_view, volumetric_fog) in view_targets.iter() {
-        let world_from_view = extracted_view.world_from_view.compute_matrix();
-
-        let mut view_fog_volumes = vec![];
-
-        for ((_, fog_volume, _), local_from_world) in
-            fog_volumes.iter().zip(local_from_world_matrices.iter())
-        {
-            // Calculate the transforms to and from 1×1×1 local space.
-            let local_from_view = *local_from_world * world_from_view;
-            let view_from_local = local_from_view.inverse();
-
-            // Determine whether the camera is inside or outside the volume, and
-            // calculate the clip space transform.
-            let interior = camera_is_inside_fog_volume(&local_from_view);
-            let hull_clip_from_local = calculate_fog_volume_clip_from_local_transforms(
-                interior,
-                &extracted_view.clip_from_view,
-                &view_from_local,
-            );
-
-            // Calculate the radius of the sphere that bounds the fog volume.
-            let bounding_radius = (Mat3A::from_mat4(view_from_local) * Vec3A::splat(0.5)).length();
-
-            // Write out our uniform.
-            let uniform_buffer_offset = writer.write(&VolumetricFogUniform {
-                clip_from_local: hull_clip_from_local,
-                uvw_from_world: UVW_FROM_LOCAL * *local_from_world,
-                far_planes: get_far_planes(&view_from_local),
-                fog_color: fog_volume.fog_color.to_linear().to_vec3(),
-                light_tint: fog_volume.light_tint.to_linear().to_vec3(),
-                ambient_color: volumetric_fog.ambient_color.to_linear().to_vec3(),
-                ambient_intensity: volumetric_fog.ambient_intensity,
-                step_count: volumetric_fog.step_count,
-                bounding_radius,
-                absorption: fog_volume.absorption,
-                scattering: fog_volume.scattering,
-                density: fog_volume.density_factor,
-                density_texture_offset: fog_volume.density_texture_offset,
-                scattering_asymmetry: fog_volume.scattering_asymmetry,
-                light_intensity: fog_volume.light_intensity,
-                jitter_strength: volumetric_fog.jitter,
-            });
-
-            view_fog_volumes.push(ViewFogVolume {
-                uniform_buffer_offset,
-                exterior: !interior,
-                density_texture: fog_volume.density_texture.as_ref().map(Handle::id),
-            });
-        }
-
-        commands
-            .entity(view_entity)
-            .insert(ViewVolumetricFog(view_fog_volumes));
-    }
-}
-
-/// A system that marks all view depth textures as readable in shaders.
-///
-/// The volumetric lighting pass needs to do this, and it doesn't happen by
-/// default.
-pub fn prepare_view_depth_textures_for_volumetric_fog(
-    mut view_targets: Query<&mut Camera3d>,
-    fog_volumes: Query<&VolumetricFog>,
-) {
-    if fog_volumes.is_empty() {
-        return;
-    }
-
-    for mut camera in view_targets.iter_mut() {
-        camera.depth_texture_usages.0 |= TextureUsages::TEXTURE_BINDING.bits();
-    }
-}
-
-fn get_far_planes(view_from_local: &Mat4) -> [Vec4; 3] {
-    let (mut far_planes, mut next_index) = ([Vec4::ZERO; 3], 0);
-    let view_from_normal_local = Mat3A::from_mat4(*view_from_local);
-
-    for &local_normal in &[
-        Vec3A::X,
-        Vec3A::NEG_X,
-        Vec3A::Y,
-        Vec3A::NEG_Y,
-        Vec3A::Z,
-        Vec3A::NEG_Z,
-    ] {
-        let view_normal = (view_from_normal_local * local_normal).normalize_or_zero();
-        if view_normal.z <= 0.0 {
-            continue;
-        }
-
-        let view_position = *view_from_local * (-local_normal * 0.5).extend(1.0);
-        let plane_coords = view_normal.extend(-view_normal.dot(view_position.xyz().into()));
-
-        far_planes[next_index] = plane_coords;
-        next_index += 1;
-        if next_index == far_planes.len() {
-            continue;
-        }
-    }
-
-    far_planes
-}
-
 impl VolumetricFogBindGroupLayoutKey {
     /// Creates an appropriate debug description for the bind group layout with
     /// these flags.
@@ -839,37 +561,4 @@ impl VolumetricFogBindGroupLayoutKey {
                 .join(", ")
         )
     }
-}
-
-/// Given the transform from the view to the 1×1×1 cube in local fog volume
-/// space, returns true if the camera is inside the volume.
-fn camera_is_inside_fog_volume(local_from_view: &Mat4) -> bool {
-    Vec3A::from(local_from_view.col(3).xyz())
-        .abs()
-        .cmple(Vec3A::splat(0.5))
-        .all()
-}
-
-/// Given the local transforms, returns the matrix that transforms model space
-/// to clip space.
-fn calculate_fog_volume_clip_from_local_transforms(
-    interior: bool,
-    clip_from_view: &Mat4,
-    view_from_local: &Mat4,
-) -> Mat4 {
-    if !interior {
-        return *clip_from_view * *view_from_local;
-    }
-
-    // If the camera is inside the fog volume, then we'll be rendering a full
-    // screen quad. The shader will start its raymarch at the fragment depth
-    // value, however, so we need to make sure that the depth of the full screen
-    // quad is at the near clip plane `z_near`.
-    let z_near = clip_from_view.w_axis[2];
-    Mat4::from_cols(
-        vec4(z_near, 0.0, 0.0, 0.0),
-        vec4(0.0, z_near, 0.0, 0.0),
-        vec4(0.0, 0.0, 0.0, 0.0),
-        vec4(0.0, 0.0, z_near, z_near),
-    )
 }
